requirements.txt
viam-sdk==0.13.3
paho-mqtt==2.1.0
